Add CompositePriorDict
The best solution I could find for mixed grid / stochastic sampling is the CompositePriorDict.
Essentially it's the old prior dict, with a rewritten .sample() method generating a grid in requested parameters.
This only accepts sample sizes which are integer multiples of the grid size (i.e. for a 3x3 .sample(size=N) must have N=9j)